fix: service bindings
The previous implementation of service bindings was panicing
because we didn't properly instantiated the
CloudControllerServiceBindingRepository, and i18n.T was nil.

With the current implementation, there is no need to instantiate
the command that way. We now solely rely on the `commandregistry`
package and the `api` one as the CLI does.

In addition to this, a smoke test was added
(TestGetServiceBindRepository). This lets us test (with a CF
connection) if the service binding works as expected.
Note: this test assumes that the service is not bound yet,
therefore an error when the service is already bound is
normal.