fix(platform/deploy): only deploy the app after service binding
Due to a mistake in the operations order, service binding
was happening _after_ the deployment has started, causing
some weird issue if you were relying on the VCAP_SERVICES
environment variable to be set-up correctly.

This commit fixes the issue, now the order of operations seems
correct and the VCAP_SERVICES environment variable is set
before the build is created.